Global Styles Render with SSR
