+++ src/client/Home.js
@@ -19,6 +19,7 @@ class Home extends React.Component {
     }
 
     render() {
+
         return (
             <div className="jarombek-main">
                 <Helmet>
