+++ src/client/styles/mixins/dyslexie.scss
@@ -1,3 +1,6 @@
+/**
+ * A mixin that sets the font to dyslexie
+ */
 @mixin dyslexie {
   background-color: transparent !important;
   font-family: 'Dyslexie', Helvetica, sans-serif !important;
