+++ src/client/styles/mixins/fantasque.scss
@@ -1,3 +1,6 @@
+/**
+ * A mixin that sets the font to fantasque bold
+ */
 @mixin fantasque-bold {
     font: {
       family: 'Fantasque-Bold', Helvetica, sans-serif !important;
